test: separated each test with append every functions at lodash
